Correctifs logs
Former-commit-id: 8610fc9c157b87abdd2557e58ba35a933c23a1cb

+++ htdocs/bimpcommercial/objects/ObjectLine.class.php
@@ -4739,10 +4739,14 @@ class ObjectLine extends BimpObject
 
             if (!$this->isRemisable()) {
                 $remises = $this->getRemises();
-                foreach ($remises as $remise) {
-                    $del_warnings = array();
-                    $remise->delete($del_warnings, true);
+
+                if (is_array($remises)) {
+                    foreach ($remises as $remise) {
+                        $del_warnings = array();
+                        $remise->delete($del_warnings, true);
+                    }
                 }
+
                 unset($this->remises);
                 $this->remises = null;
             }
